Moved test cases to proper folder
