+++ tests/Unit/ErrorPageRenderingServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Unit;
 
 use App\Services\ErrorPageRenderingService;
 use Tests\TestCase;
