+++ tests/Unit/InquiryRequestTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Unit;
 
 use App\Http\Requests\InquiryRequest;
 use Tests\TestCase;
