+++ tests/Unit/PageRenderingServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Unit;
 
 use App\Services\PageRenderingService;
 use Tests\TestCase;
